restructure styling to prevent bleed
The CSS outside the client is now pulled to within the ShadowDOM. Since
some rules were required on the outside, POLAR now constructs an
additional wrapper to itself and adds a style tag containing the
required construct styles.

In this version, fontawesome icons are broken. This is due to a bug in
browsers. E.g. the Chromium issue:
https://bugs.chromium.org/p/chromium/issues/detail?id=336876

A workaround to this will be added in a following commit.

Relates: #50