add log to note about render issue resolution
OL will `console.warn` on a rendering issue where height and/or width
of the map are zero. Since the map client retries rendering until it
works, this may come across as off, since the `console.warn` will not
match the fact that the map is rendered eventually.

A `console.log` has been added to correct this information.

Relates: #50