add new `stylePath` key to MapConfiguration
Instead of providing the styles via DOM, they can now be provided via
configuration by adding a relative path that leads to the CSS file. This
is a more clean solution since it prevents a flash of misstyled content.

It is also required for SPA since modifying the DOM on map creation will
lead to misbehaviour on subsequent calls; that is, the DOM provision is
no longer available then.

The previous DOM-based method of style provision has been deprecated,
but is still used in DISH and supported for the rest of the current
major version in favour of AfM clients.

Relates: #50