fix fontawesome icon rendering issue
The ShadowDOM is not completely working since @font-face instructions do
not work correctly. This is e.g. documented in
https://bugs.chromium.org/p/chromium/issues/detail?id=336876
. As a hack to circumvent this issue, POLAR will now add fontawesome
styles to the Light/Root DOM. This may become an issue with conflicting
fontawesome versions in the wrapping application; POLAR's behaviour
regarding this can be disabled via configuration as a first attempt to
resolve this. Since that may or may not work, further changes may be
required – maybe browsers fix this in time before the use case arises.

Else, we'll pay that troll toll when we get there.

Resolves: #50