update DISH to new AddressSearch mechanism
Making the plugin AddressSearch SPA ready changed some undocumented
properties of the plugin. DISH was previously using the store directly
by importing it from the package. Now, this is no longer possible, and
instead the executing method's context is set to the created store.

A NOTE has been left in the CHANGELOG since this is relevant to the next
release's dependency versions.

Relates: #66