Fix typescript setup
The package @polar/lib-custom-types and the folder vue-shims are
fundamentally different in its usage. The vue-shims are (probably) used
to enhance the typescript experience of vue files while the package
holds explicit types.

Now, all types in the configured typeRoots are used.

Also added the parameter isolatedModules as described in the vite docs:
https://vitejs.dev/guide/features.html#typescript-compiler-options