change geometry tooltip display to primary names
Previously, all names of all geometries were shown as a list. Now, only
primary names of an object are used, and if an object holds multiple
primary names, they are listed in a single list item.

There seem to be data errors where objects are without names. They have
been reported. For the time being, '???' is displayed for those items,
in the same fashion as it's done in the Küstengazetteer.