Move duplicated prop isHorizontal to core as getter deviceIsHorizontal
Also, remove redundant bottom option from tooltips of the IconMenu. They
are only shown if they are on a large device thus they are (currently)
always on the left side of the icon.

If the placement of the IconMenu changes in the future, this behaviour
needs to be configured in sync with the other css classes that are set
in assumption of the current positioning.