update eslint to @9
This updates eslint and our related dependencies to the newest major
version. Not everything is working yet, but it is linting again, and
some improvements have been reached in the process.

However, now builds are broken, which is probably due to the required
changes in the tsconfig.json, and the tsdoc linting rules have not yet
been working again. Hence, this is a WIP commit.