Add first draft for selecting multiple features
Also, add new configuration parameter directSelect to be able to select
features directly from GFI and be able to select multiple features with
multiple clicks.

To be done is the combination with useExtendedMasterportalapiMarkers,
information for the user in the FeatureList-UI, adding Changelog entries
and testing it in both DISH and Meldemichel.