Add additional mode 'none' and internal variable to select interaction
By adding these two, it is now possible to measure lengths and areas
and not triggering the pins plugin or the directSelect of gfi.
If one wants to use more complex interactions, they may remove the
measure interactions by selecting "No interaction".
This still leaves room open for draw and measure to work together, but
this may be a wanted situation. If not, this can be removed by simply
toggling the respective plugin.