Adjust implementation of the MoveHandle in GFI to properly work again
After removing several computed values and props in #131 the mutation to
set the gfi in the MoveHandle through the watcher was no longer
triggering as the values of moveHandleProperties were always the same.

Now, the GFI window is displayed again on mobile devices when it's
rendered independently and the closeIcon is now again fa-angles-right
when featureList is configured.