Refactor map initialization
Add defaults not in the store but in the createMap function itself so
that they are actually registered.
Also, call initializeLayerList in the component so the callback can be
used directly. This also means, that the layerConf configuration
parameter now has been replaced by the serviceRegister argument of the
createMap function.