chore(release): version 3.0.0-alpha.0
This version provides `@polar/polar`, which supersedes the packages
`@polar/core`, `@polar/lib-X` and `@polar/plugin-X`.

The core is already working, but there are still a lot of plugins
needing migration. The not-yet-migrated code is left in the `vue2`
folder until it's migrated.

This version was not heavily tested yet, so use with care and not in
production.

Starting with the next version, POLAR uses conventional commits for
changelog generation.